+++ app/Http/Controllers/FavoriteArtistController.php
@@ -2,15 +2,14 @@
 
 namespace App\Http\Controllers;
 
-
+use App\Http\Requests\UpdateArtistRequest;
 use App\Models\FavoriteArtist;
 use Exception;
-use Illuminate\Http\Request;
 use Inertia\Inertia;
 use Inertia\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\RedirectResponse;
-
+use Illuminate\Http\Request;
 
 class FavoriteArtistController extends Controller
 {
@@ -20,7 +19,10 @@ class FavoriteArtistController extends Controller
      */
     public function index(): Response
     {
-        $favArtists = FavoriteArtist::where("userId",  Auth::id())->orderBy('created_at', 'desc')->get();
+       $this->authorize('view', FavoriteArtist::class);
+
+        $favArtists = FavoriteArtist::where("userId",  Auth::id())
+        ->orderBy('created_at', 'desc')->paginate(5);
         return Inertia::render("Artists", ["favoriteArtists" => $favArtists]);
     }
 
@@ -29,16 +31,13 @@ class FavoriteArtistController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
-
+        $this->authorize('create', FavoriteArtist::class);
         try {
-            $request->validate([
-                "name" => "required|string|unique:favorite_artists"
-            ]);
-
-            FavoriteArtist::create(array_merge($request->all(), ["userId" => $request->user()->id]));
+            $data = array_merge($request->all(), ["userId" => Auth::id()]);
+            FavoriteArtist::create($data);
             return redirect("/artist");
         } catch (Exception $e) {
-            return redirect("artist");
+            return redirect("/artist");
         }
     }
 
@@ -53,19 +52,12 @@ class FavoriteArtistController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, FavoriteArtist $favoriteArtist): RedirectResponse
+    public function update(UpdateArtistRequest $request, FavoriteArtist $favoriteArtist): RedirectResponse
     {
-
-        $validated = $request->validate([
-            "name" => "required|string",
-            "listeners" => "required",
-            "mbid" => "nullable",
-            "streamable" => "required",
-            "url" => "url:http,https|required",
-
-        ]);
-
+        $validated = $request->validated();
         $artist = FavoriteArtist::find($request->route('artist'));
+       
+        $this->authorize("update", $artist);
         $artist->update($validated);
         return redirect()->back();
     }
@@ -76,6 +68,7 @@ class FavoriteArtistController extends Controller
     public function destroy(int $artistId): RedirectResponse
     {
         $artist = FavoriteArtist::find($artistId);
+        $this->authorize("delete", $artist);
         $artist->delete();
         return redirect()->back();
     }
